# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Driveway Estimator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EstimateInput {
  customerName: string;
  phone: string;
  email: string;
  address: string;
  /** Estimated square footage of the driveway */
  squareFootage: number;
  /** Whether trees/obstructions were detected */
  hasTreeObstruction?: boolean;
  /** Sealing price before crack fill */
  basePrice: number;
  /**
     * Additional crack fill price if selected
     * @nullable
     */
  crackFillPrice?: number | null;
  /** Total estimated price */
  totalPrice: number;
  /** Whether crack fill was selected */
  hasCrackFill: boolean;
  /**
     * Optional notes
     * @nullable
     */
  notes?: string | null;
}

export interface Estimate {
  id: number;
  customerName: string;
  phone: string;
  email: string;
  address: string;
  squareFootage: number;
  hasTreeObstruction?: boolean;
  basePrice: number;
  /** @nullable */
  crackFillPrice?: number | null;
  totalPrice: number;
  hasCrackFill: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface QrCode {
  /** Base64 encoded QR code image */
  dataUrl: string;
  /** The URL encoded in the QR code */
  url: string;
}

export interface ErrorResponse {
  error: string;
}