# lib/api-zod/src/generated/types/estimate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Driveway Estimator API
 * OpenAPI spec version: 0.1.0
 */

export interface Estimate {
  id: number;
  customerName: string;
  phone: string;
  email: string;
  address: string;
  squareFootage: number;
  hasTreeObstruction?: boolean;
  basePrice: number;
  /** @nullable */
  crackFillPrice?: number | null;
  totalPrice: number;
  hasCrackFill: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}