# lib/api-zod/src/generated/types/estimateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Driveway Estimator API
 * OpenAPI spec version: 0.1.0
 */

export interface EstimateInput {
  customerName: string;
  phone: string;
  email: string;
  address: string;
  /** Estimated square footage of the driveway */
  squareFootage: number;
  /** Whether trees/obstructions were detected */
  hasTreeObstruction?: boolean;
  /** Sealing price before crack fill */
  basePrice: number;
  /**
     * Additional crack fill price if selected
     * @nullable
     */
  crackFillPrice?: number | null;
  /** Total estimated price */
  totalPrice: number;
  /** Whether crack fill was selected */
  hasCrackFill: boolean;
  /**
     * Optional notes
     * @nullable
     */
  notes?: string | null;
}