# lib/api-zod/src/generated/types/qrCode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Driveway Estimator API
 * OpenAPI spec version: 0.1.0
 */

export interface QrCode {
  /** Base64 encoded QR code image */
  dataUrl: string;
  /** The URL encoded in the QR code */
  url: string;
}